fix: render in 32bits #1214

+++ src/framework/graphics/texturemanager.cpp
@@ -198,7 +198,7 @@ const Matrix3* TextureManager::getMatrixById(uint16_t id) {
 }
 
 uint16_t TextureManager::getMatrixId(const Size& size, bool upsidedown) {
-    const size_t hash = (static_cast<uint64_t>(size.height()) << 33) | (static_cast<uint64_t>(size.width()) << 1) | (upsidedown ? 1 : 0);
+    const uint64_t hash = (static_cast<uint64_t>(size.height()) << 33) | (static_cast<uint64_t>(size.width()) << 1) | (upsidedown ? 1 : 0);
     auto it = m_matrixCache.indexMap.find(hash);
     if (it != m_matrixCache.indexMap.end()) {
         return it->second;
