+++ src/framework/graphics/texturemanager.h
@@ -52,7 +52,7 @@ private:
 
     struct
     {
-        std::unordered_map<size_t, uint16_t> indexMap;
+        std::unordered_map<uint64_t, uint16_t> indexMap;
         std::vector<std::unique_ptr<Matrix3>> objects;
     } m_matrixCache;
